+++ src/commands/handlers/event/remind.ts
@@ -5,7 +5,7 @@
  */
 
 import { type CacheType, type ChatInputCommandInteraction, type Client, MessageFlags } from 'discord.js';
-import eventLang from '../../../lang/event.json';
+import eventLang from '../../../lang/en/event.json';
 import { EventConfig } from '../../../typeorm/entities/event/EventConfig';
 import { EventReminder } from '../../../typeorm/entities/event/EventReminder';
 import { enhancedLogger, guardAdmin, LANGF, LogCategory } from '../../../utils';